Add files via upload
I have uploaded two code files: Assignment2 and Assignment2_updated. In Assignment2_updated, I have removed the "New_Price" column and have not included it in the subsequent four tasks. Conversely, in the Assignment2 file, I deleted the "New_Price" column but mistakenly included it in the remaining four tasks.